chore: use transaction list item component
